botsy/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
schedule>=1.2.0
lxml>=4.9.0

# News & Media
feedparser>=6.0.0
newspaper3k>=0.2.8
trafilatura>=1.6.0

# Research & Academia
arxiv>=1.4.0
scholarly>=1.7.0
crossref-commons>=0.0.7

# Finance & Markets
yfinance>=0.2.0
alpha-vantage>=2.3.0
pandas-datareader>=0.10.0

# Technology & Software
PyGithub>=1.59.0
requests-html>=0.10.0

# Social Media
tweepy>=4.14.0
praw>=7.7.0
instaloader>=4.9.0

# Government & Legal
requests-cache>=1.1.0

# Weather & Environment
pyowm>=3.3.0

# Sports & Entertainment
espn-api>=0.3.0

# E-commerce & Reviews
selenium>=4.11.0

# Health & Medical
biopython>=1.81

# Additional specialized libraries
# GitHub API
pygithub>=1.59.0

# Academic research
semanticscholar>=0.3.0

# Data processing
polars>=0.19.0
dask>=2023.9.0

# Web scraping enhancement
scrapy>=2.11.0
playwright>=1.37.0
requests-html>=0.10.0

# Text processing
nltk>=3.8.0
textblob>=0.17.0
sumy>=0.11.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0

# Database connectivity
sqlalchemy>=2.0.0
pymongo>=4.5.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
colorama>=0.4.6
tqdm>=4.65.0
click>=8.1.0
rich>=13.5.0

# Optional advanced dependencies (commented out by default)
# Advanced ML/NLP
# transformers>=4.33.0
# torch>=2.0.0
# spacy>=3.6.0

# Advanced data processing  
# apache-airflow>=2.7.0
# celery>=5.3.0

# Cloud integrations
# boto3>=1.28.0
# google-cloud-storage>=2.10.0
# azure-storage-blob>=12.17.0